Updating shop with dynamic ratings
We will need to think about validations of the dynamic ratings. Opportunity to refactor shop with numericality instead of regex.

We will need to do more useful data manipulation than just replacing the value. We want averages of the most recent 5 dynamic ratings to start, but seed it as an admin.

Potential opportunity to refactor private method to use only instance variables and no parameters. Using conditionals to ensure that blank dynamic ratings don't break shop validations.

We rejected drying out the test with metaprogramming because it seemed too hard. Test has only tested that the value has changed, not the algorithm itself.

https://app.asana.com/0/11389918903342/11518383302367